producer: Initial WebsiteCheck class definition
This patch creates the monitor python module with just one file, the
producer.py. In this file there's a new class to define what's a website
check and the behavior needed.

Right now the WebsiteCheck class just wraps the requests.get method and
stores the information in the class, this is an abstractions that will
allow to change the request method in the future and simplifies the API
to create and manage website checks.

A new Pipfile with dependencies has been added to make it easy to
replicate the python virtualenv.