producer: Add command line management
Adds the command line management for the producer module so it can be
called from the command line with `python -m`.

The --monitor option provides a way to launch this program as a daemon
so it creates a loop and do a check for this website every N seconds.

Some terminal colors has been added, red for errors and blue for info.
Maybe it could be interesting to use some existing python module for
this, like the termcolor [1], but I think that it's not needed a new
dependency just for this.

[1] https://pypi.org/project/termcolor/